# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CUI API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/sessions": {
      "get": {
        "summary": "List sessions",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 100
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of session summaries with current running state.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "origin": {
                            "type": "string",
                            "enum": ["chat", "shell"]
                          },
                          "workspace": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "summary": {
                            "type": "string"
                          },
                          "doneAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "queuedPrompts": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string"
                                },
                                "mode": {
                                  "type": "string",
                                  "enum": ["chat", "shell"]
                                },
                                "prompt": {
                                  "type": "string"
                                },
                                "createdAt": {
                                  "type": "string",
                                  "format": "date-time"
                                }
                              },
                              "required": ["id", "mode", "prompt", "createdAt"]
                            }
                          },
                          "currentRound": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "gitBranch": {
                            "type": "string"
                          },
                          "isRunning": {
                            "type": "boolean"
                          },
                          "runningRunId": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "workspace",
                          "title",
                          "createdAt",
                          "updatedAt",
                          "currentRound",
                          "isRunning"
                        ]
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer",
                          "exclusiveMinimum": 0
                        },
                        "pageSize": {
                          "type": "integer",
                          "exclusiveMinimum": 0
                        },
                        "total": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "totalPages": {
                          "type": "integer",
                          "exclusiveMinimum": 0
                        },
                        "hasPreviousPage": {
                          "type": "boolean"
                        },
                        "hasNextPage": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "page",
                        "pageSize",
                        "total",
                        "totalPages",
                        "hasPreviousPage",
                        "hasNextPage"
                      ]
                    }
                  },
                  "required": ["sessions", "pagination"]
                }
              }
            }
          },
          "400": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspace": {
                    "type": "string",
                    "minLength": 1
                  },
                  "origin": {
                    "type": "string",
                    "enum": ["chat", "shell"]
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": ["workspace"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The session was created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "session": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "origin": {
                          "type": "string",
                          "enum": ["chat", "shell"]
                        },
                        "workspace": {
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        },
                        "summary": {
                          "type": "string"
                        },
                        "doneAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "messages": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "role": {
                                "type": "string",
                                "enum": ["assistant", "user"]
                              },
                              "kind": {
                                "type": "string",
                                "enum": ["response", "trace"]
                              },
                              "round": {
                                "type": "integer",
                                "exclusiveMinimum": 0
                              },
                              "content": {
                                "type": "string"
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            },
                            "required": ["id", "role", "content", "createdAt"]
                          }
                        },
                        "rounds": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "round": {
                                "type": "integer",
                                "exclusiveMinimum": 0
                              },
                              "hasChanges": {
                                "type": "boolean"
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "atomicReviewStatus": {
                                "anyOf": [
                                  {
                                    "type": "string",
                                    "enum": ["ready"]
                                  },
                                  {
                                    "type": "string",
                                    "enum": ["failed"]
                                  }
                                ]
                              }
                            },
                            "required": ["round", "hasChanges", "createdAt"]
                          }
                        },
                        "queuedPrompts": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "mode": {
                                "type": "string",
                                "enum": ["chat", "shell"]
                              },
                              "prompt": {
                                "type": "string"
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            },
                            "required": ["id", "mode", "prompt", "createdAt"]
                          }
                        },
                        "currentRound": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "gitBranch": {
                          "type": "string"
                        },
                        "isRunning": {
                          "type": "boolean"
                        },
                        "runningRunId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "workspace",
                        "title",
                        "createdAt",
                        "updatedAt",
                        "messages",
                        "currentRound",
                        "isRunning"
                      ]
                    }
                  },
                  "required": ["session"]
                }
              }
            }
          },
          "400": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sessions/{sessionId}/rounds/{round}/review": {
      "get": {
        "summary": "Get a round review",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "sessionId",
            "in": "path"
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "required": true,
            "name": "round",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested round review.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "review": {
                      "type": "object",
                      "properties": {
                        "round": {
                          "type": "integer",
                          "exclusiveMinimum": 0
                        },
                        "baseCommit": {
                          "type": "string"
                        },
                        "beforeDiff": {
                          "type": "string"
                        },
                        "afterDiff": {
                          "type": "string"
                        },
                        "diff": {
                          "type": "string"
                        },
                        "hasChanges": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "atomicReview": {
                          "oneOf": [
                            {
                              "type": "object",
                              "properties": {
                                "status": {
                                  "type": "string",
                                  "enum": ["ready"]
                                },
                                "generatedAt": {
                                  "type": "string",
                                  "format": "date-time"
                                },
                                "analysisSessionId": {
                                  "type": "string"
                                },
                                "items": {
                                  "type": "array",
                                  "items": {
                                    "type": "object",
                                    "properties": {
                                      "id": {
                                        "type": "string"
                                      },
                                      "order": {
                                        "type": "integer"
                                      },
                                      "capabilityType": {
                                        "anyOf": [
                                          {
                                            "type": "number",
                                            "enum": [0]
                                          },
                                          {
                                            "type": "number",
                                            "enum": [1]
                                          },
                                          {
                                            "type": "number",
                                            "enum": [2]
                                          },
                                          {
                                            "type": "number",
                                            "enum": [3]
                                          },
                                          {
                                            "type": "number",
                                            "enum": [5]
                                          }
                                        ]
                                      },
                                      "capabilityLabel": {
                                        "type": "string"
                                      },
                                      "title": {
                                        "type": "string"
                                      },
                                      "intent": {
                                        "type": "string"
                                      },
                                      "files": {
                                        "type": "array",
                                        "items": {
                                          "type": "string"
                                        }
                                      },
                                      "diff": {
                                        "type": "string"
                                      },
                                      "outputJson": {
                                        "type": "object",
                                        "additionalProperties": {}
                                      }
                                    },
                                    "required": [
                                      "id",
                                      "order",
                                      "capabilityType",
                                      "capabilityLabel",
                                      "title",
                                      "intent",
                                      "files",
                                      "diff",
                                      "outputJson"
                                    ]
                                  }
                                },
                                "rawResponse": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "status",
                                "generatedAt",
                                "analysisSessionId",
                                "items",
                                "rawResponse"
                              ]
                            },
                            {
                              "type": "object",
                              "properties": {
                                "status": {
                                  "type": "string",
                                  "enum": ["failed"]
                                },
                                "generatedAt": {
                                  "type": "string",
                                  "format": "date-time"
                                },
                                "error": {
                                  "type": "string"
                                },
                                "rawResponse": {
                                  "type": "string"
                                }
                              },
                              "required": ["status", "generatedAt", "error"]
                            }
                          ]
                        }
                      },
                      "required": [
                        "round",
                        "beforeDiff",
                        "afterDiff",
                        "diff",
                        "hasChanges",
                        "createdAt"
                      ]
                    }
                  },
                  "required": ["review"]
                }
              }
            }
          },
          "400": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "404": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}